renamed: activation -> annotation
