+++ src/annotation/tagged.ts
@@ -1,6 +1,6 @@
 ///<reference path="../interfaces/interfaces.d.ts" />
 
-import Metadata from "../activation/metadata";
+import Metadata from "../planning/metadata";
 import { tagParameter } from "./decorator_utils";
 
 // Used to add custom metadata which is used to resolve metadata-based contextual bindings.
