+++ src/interfaces/planning/context.d.ts
@@ -1,9 +1,9 @@
 interface IContext {
 
-        /// Gets the kernel that is driving the activation.
+        /// Gets the kernel that is driving the resolution.
         kernel: IKernel;
 
-        /// Gets or sets the activation plan.
+        /// Gets or sets the resolution plan.
         plan: IPlan;
 
         addPlan(plan: IPlan): void;
