+++ src/inversify.ts
@@ -6,11 +6,11 @@
 // The Inversify main file, the library entry point.
 
 import Kernel from "./kernel/kernel";
-import inject from "./activation/inject";
-import tagged from "./activation/tagged";
-import named from "./activation/named";
-import paramNames from "./activation/paramnames";
-import { decorate } from "./activation/decorator_utils";
+import inject from "./annotation/inject";
+import tagged from "./annotation/tagged";
+import named from "./annotation/named";
+import paramNames from "./annotation/paramnames";
+import { decorate } from "./annotation/decorator_utils";
 
 export { Kernel };
 export { decorate };
