+++ test/annotation/named.test.ts
@@ -4,8 +4,8 @@ declare function __param(paramIndex: number, decorator: ParameterDecorator): Cla
 ///<reference path="../../src/interfaces/interfaces.d.ts" />
 
 import { expect } from "chai";
-import { decorate } from "../../src/activation/decorator_utils";
-import named from "../../src/activation/named";
+import { decorate } from "../../src/annotation/decorator_utils";
+import named from "../../src/annotation/named";
 import * as METADATA_KEY from "../../src/constants/metadata_keys";
 
 interface IWeapon {}
