+++ test/resolution/resolver.test.ts
@@ -8,10 +8,10 @@ import Kernel from "../../src/kernel/kernel";
 import Request from "../../src/planning/request";
 import Plan from "../../src/planning/plan";
 import Target from "../../src/planning/target";
-import inject from "../../src/activation/inject";
-import tagged from "../../src/activation/tagged";
-import named from "../../src/activation/named";
-import paramNames from "../../src/activation/paramnames";
+import inject from "../../src/annotation/inject";
+import tagged from "../../src/annotation/tagged";
+import named from "../../src/annotation/named";
+import paramNames from "../../src/annotation/paramnames";
 import * as ERROR_MSGS from "../../src/constants/error_msgs";
 import BindingType from "../../src/bindings/binding_type";
 import * as Proxy from "harmony-proxy";
